==> python_codes/1_data_gathering_women_new.ipynb <==
# Web scraping to gather data: Women's New Collection

> - Using the smallable website, men's and women's products will be scraped to prepare for creating a clothing and accessories recommendation.

# try to webscrap smallable
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import requests
from time import sleep
import random

# women's new collection
w_new_pages = range(1,7)

w_new = f"https://www.smallable.com/en/fashion/adult/women?_collection=new-collection&_page="

# dataframe where all results will be stored
womens_new = pd.DataFrame()

# iteration for webscrapping multiple pages
for page in w_new_pages: #change pages on the link
    r = requests.get(f"https://www.smallable.com/en/fashion/adult/women?_collection=new-collection&_page={page}")
    print(f'Response for page {page} :', r.status_code)
    soup = BeautifulSoup(r.content, 'html.parser')
    
    #empty lists
    link = []
    img = []
    tags = []
    brand = []
    product_color = []
    price_euro = []

    #scrapping infos from site
    for i in soup.find_all('a', attrs={"class": "ProductCard_content__fBfLV"}):
        link.append(i.get("href")) #link to product
        img.append(i.select('img')[1].get("src")) # photo link
        tags.append(i.find("div", attrs = {"class" : "ProductTags_container__3cc_M"}).get_text(strip = True, separator = "|"))

    for i in soup.find_all('ul', attrs={"class": "ProductCard_attr__2e2YT"}):
        brand.append(i.select("li")[0].get_text(strip = True)) #brand name
        product_color.append(i.select("li")[1].get_text(strip = True)) #product | color
        price_euro.append(i.select("li")[2].get_text(strip = True, separator = "|")) #price
    
    new_df = pd.DataFrame(list(zip(link, img, tags, brand, product_color, price_euro)),columns=("link", "img", "tags", "brand", "product_color", "price_euro"))
    # store everything in the previously prepared df and concatenate new results from new webpages
    womens_new = pd.concat([womens_new, new_df])
    
    wait_time = random.randint(2,7)
    print("I will sleep for " + str(wait_time) + " seconds.")
    sleep(wait_time) #Script will be stopped for a period of time ranging from 2 to 7 seconds

womens_new.to_csv('womens_new.csv', index=False)

womens_new.duplicated().sum()

womens_new

# code adapted from Cormac and Isi
def request_link(link):
    try:
        request = requests.get(link)
        request.raise_for_status()  # returns an HTTPError if the response is not OK
        print("Success! Response code", request.status_code) 
    except requests.exceptions.HTTPError as err:
        if request.status_code == 404:
            print("404: Oops, sorry we can't find that page!")
        else:
            print("The error code is", err.args[0]) # look up the 1st argument from HTTPError 
    return request

w = request_link(w_new)

soup = BeautifulSoup(w.content, "html.parser")

for item in soup.find_all('a', attrs={"class": "ProductCard_content__fBfLV"}):
        print(item.get("href")) #link to product
        print(item.select('img')[1].get("src")) # photo link
        print(item.find("div", attrs = {"class" : "ProductTags_container__3cc_M"}).get_text(strip = True, separator = "|"))

for i in soup.find_all('ul', attrs={"class": "ProductCard_attr__2e2YT"}):
        print(i.select("li")[0].get_text(strip = True)) #brand name
        print(i.select("li")[1].get_text(strip = True)) #product | color
        print(i.select("li")[2].get_text(strip = True, separator = "|")) #price
