==> python_codes/2_data_cleaning_updated.ipynb <==
# CLEANING WEB SCRAPED DATASETS

## IMPORT LIBRARIES

import pandas as pd
import numpy as np
import re

## LOAD DATASETS

# women's datasets
women_ss = pd.read_csv(r'./../data/web_scraped/womens_ss.csv')
women_aw = pd.read_csv(r'./../data/web_scraped/womens_aw.csv')
women_new = pd.read_csv(r'./../data/web_scraped/womens_new.csv')

women_ss.sample(10)

# men's dataset
men_ss = pd.read_csv(r'./../data/web_scraped/mens_ss.csv')
men_aw = pd.read_csv(r'./../data/web_scraped/mens_aw.csv')
men_new = pd.read_csv(r'./../data/web_scraped/mens_new.csv')

men_ss.sample(10)

## FUNCTIONS FOR CLEANING SPECIFIC COLUMNS

> #### After looking at the different dataframes, the columns to clean are as follows:
>   - Tags : Sometimes has multiple values per row. Most interesting of them is the 'greenable' tag which implies the product has passed 2 criteria:
> 
>          1. This item has been made from natural materials, ingredients or fibres.
>          2. This item’s packaging has been made from recycled or biodegradable materials (which decompose into natural components found in nature).
> 
>   - Product_color : has 2 different information so they will be split into 2 columns corresponding to product_detail and color.
>   
>          - The product detail also contain the type of product (i.e. dress, coat, pullover, etc...). This could be extracted to create a label which will be useful when checking the performance of the image classification later on.
>          - There are a lot of different color names provided. This column could also be cleaned to get only general colors but since we would not be exploiting the color column, we will forgo this step to gain time for the rest of the project.
>   
>   - Price :  will be converted to float. It is currently an object because of special characters (€, -, %).
>   
>          - We can also see some prices are discounted so these information will be transfered to another column.
>          - Also, the number of colors available for the product is mentioned in this column. This will also be removed.
> 
>   - We can also add another column to label the ss and aw collection as such. This could help filter in which collection the products to be recommended should be taken from, if needed. Since there are not a lot of products in the new collection, this dataset will not be used for both men and women's collection.

def split_prod_col(df):
    """This function cleans and splits the product_color column which contains 2 different informations.
    It will return a new dataframe which we could use with the next cleaning process."""
    
    # Separate the model column into two empty lists
    product = []
    color = []
    
    # split list
    split = df.product_color.apply(lambda x: x.split(' | ')) 
    
    ## get each brand and product_color value for each row
    index = range(0, len(split)) 
    for i in index:
        try:
            product.append(split[i][0] if split[i][0] else 'none') # takes first element which is the product description
            color.append(split[i][1] if split[i][0] else 'none') # takes second element which corresponds to the product color
        except IndexError:
            color.append('none')
        
    # Insert new split columns to the dataframe
    new_df = df.copy()
    new_df.insert(3, 'product', product)
    new_df.insert(4, 'color', color)
    new_df.drop("product_color", axis=1, inplace=True) # drop the original product_color column
        
    return new_df

def get_greenable(df):
    """This function will transform the values in the tags column into lowercase and extract greenable boolean values."""
    
    #empty list
    greenable = []
    
    #lowercase values in tags
    lower = df.tags.str.lower()
    
    # checking for greenable tag
    for i in lower:
        try:
            if "greenable" in i:
                greenable.append(True)
            else:
                greenable.append(False)
        except TypeError:
            greenable.append(False)
    
    df.insert(6, "greenable", greenable)
    df.drop("tags", axis=1, inplace=True)
    df.greenable = df.greenable.astype(int)
    return df

def clean_price(df):
    """This function will clean the price column such that:
    1. new column discounted price will be created.
    2. price will indicate the product's orinal price
    3. a discounted column with bolean values will indicate if the product is currently on sale
    4. colors will be discarded since we do not have access to all the colors available for each product."""
    
    # references to search for certain values
    ref_discount = r'.[-][0-9]{2}%'
    ref_colours = r'.[0-9]{1,2}[ ]colours'
    
    # removing unwanted details
    col1 =  df.price_euro.replace(ref_colours, "", regex=True).replace("[€,]", "", regex=True).replace(r'From.', "", regex=True) # remove color details and other special characters
    
    # checking which rows are discounted
    disc = col1.apply(lambda x: re.search(ref_discount, x))
    discounted = disc.apply(lambda x: 0 if x is None else 1) # new list
    
    # split to get original and discounted prices
    col2 =  col1.replace(ref_discount, "", regex=True) # remove discount details from price col
    split = col2.apply(lambda x: x.split("|"))
    
    # original and discounted price
    orig_price = split.apply(lambda x: float(x[-1]))
    disc_price = split.apply(lambda x: float(x[0])) # note that for products which are not discounted, the orginal price is indicated here
    
    # insert new columns to the dataframe and delete original price_euros column
    df['discounted'] = discounted
    df['disc_price'] = disc_price
    df['orig_price'] = orig_price
    df.drop("price_euro", axis=1, inplace=True)
    
    return df

def lowercase_df(df):
    """This function will transform all the values in the dataframe into lowercase characters and replace accented vowels to normal characters"""
    cat = df.select_dtypes(object)
    for column in cat.columns:
        char = df["product"].replace(r'[éèê]', "e", regex=True).replace("à", "a", regex=True).replace(r'[öó]', "o", regex=True).replace("ù", "u", regex=True)
        df['product'] = char
        lower = df[column].str.lower()
        df[column] = lower
    return df

## WOMEN'S COLLECTION DATASET

### SPRING-SUMMER COLLECTION

# split product_color column into 2
wss1 = split_prod_col(women_ss)

wss1.head()

# boolean column to indicate if a product is eco-friendly or not
wss2 = get_greenable(wss1)

wss2[:10]

# clean price column which has numerous details and special characters
wss3 = clean_price(wss2)

# transform values in the df into lowercase to facilitate searches
wss4 = lowercase_df(wss3)

# save to csv
wss4.to_csv("women_ss_clean.csv", index=False)

### FALL-WINTER COLLECTION

# split product_color column into 2
wfw1 = split_prod_col(women_aw)

# boolean column to indicate if a product is eco-friendly or not
wfw2 = get_greenable(wfw1)

# clean price column which has numerous details and special characters
wfw3 = clean_price(wfw2)

# transform values in the df into lowercase to facilitate searches
wfw4 = lowercase_df(wfw3)

# save to csv
wfw4.to_csv("women_fw_clean.csv", index=False)

## MEN'S COLLECTION

### SPRING-SUMMER COLLECTION

# split product_color column into 2
mss1 = split_prod_col(men_ss)

# boolean column to indicate if a product is eco-friendly or not
mss2 = get_greenable(mss1)

# clean price column which has numerous details and special characters
mss3 = clean_price(mss2)

# transform values in the df into lowercase to facilitate searches
mss4 = lowercase_df(mss3)

# save to csv
mss4.to_csv("men_ss_clean.csv", index=False)

### FALL-WINTER COLLECTION

# split product_color column into 2
mfw1 = split_prod_col(men_aw)

# boolean column to indicate if a product is eco-friendly or not
mfw2 = get_greenable(mfw1)

# clean price column which has numerous details and special characters
mfw3 = clean_price(mfw2)

# transform values in the df into lowercase to facilitate searches
mfw4 = lowercase_df(mfw3)

# save to csv
mfw4.to_csv("men_fw_clean.csv", index=False)
