==> python_codes/4_getting_product_categories.ipynb <==
# Getting Product Categories (Again through web scraping)

## IMPORT LIBRARIES

# add more details to previously scraped datasets
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import requests
from time import sleep
import random

## LOAD DATA

all = pd.read_csv(r'./../data/cleaned_data/all_collection.csv')

## TRY OUT CODE FOR WEB SCRAPING

r = requests.get("https://www.smallable.com/en/product/parapluie-pliant-camden-navy-blue-anatole-335388")
soup = BeautifulSoup(r.content, 'html.parser')       

product_category = []

category = soup.find('nav', attrs={"class": "BreadCrumb_breadcrumb__3eArq BreadCrumb_breadcrumbLastItem__3KOwJ"}).select('span')[-1].get_text(strip=True)

product_category.append(category)

product_category

## WEB SCRAPING MULTIPLE PAGES

# empty list
product_category = []
index = []

# iteration for webscrapping multiple pages
for link, i in zip(all12.link, range(8317, len(all))):
    r = requests.get(f"https://www.smallable.com{link}")
    soup = BeautifulSoup(r.content, 'html.parser')
    print(f'Response for {link} :', r.status_code)
    #scraping infos from site
    #product category
    category = soup.find('nav', attrs={"class": "BreadCrumb_breadcrumb__3eArq BreadCrumb_breadcrumbLastItem__3KOwJ"}).select('span')[-1].get_text(strip=True)
    product_category.append(category)
    #index
    index.append(i)
    wait_time = random.randint(2,5)
    print("I will sleep for " + str(wait_time) + " seconds.")
    sleep(wait_time) #Script will be stopped for a period of time ranging from 2 to 6 seconds

> **NOTE: Code above had to be done at least 12 times (thus the all12 df below). It had to be stopped when the response stopped after a while and so proceeded with the next batch of dataset that has yet to have their categories.**

### ITERATIONS

all12 = all.iloc[8317:]

len(all12)

product_category

# sanity check everytime the code was interrupted
len(product_category)

# sanity check everytime the code was interrupted
len(index)

### SAVED RESULTS
- each time a batch of categories were acquired, they were saved into dataframes in order to concat everything later.

df12 = pd.DataFrame(product_category,index=index, columns=['product_category'])

df12

df12.to_csv('product_category_8317-8734.csv')
