==> python_codes/4.1_adding_product_categories.ipynb <==
## IMPORT LIBRARIES

import pandas as pd
import numpy as np
import re

## LOADING DATA

all = pd.read_csv(r'./../data/cleaned_data/all_collection.csv')

pc1 = pd.read_csv(r"./../data/product_categories/product_category_0-715.csv")
pc2 = pd.read_csv(r"./../data/product_categories/product_category_716-1205.csv")
pc3 = pd.read_csv(r"./../data/product_categories/product_category_1206-2304.csv", index_col="Unnamed: 0")

pc4 = pd.read_csv(r"./../data/product_categories/product_category_2305-3032.csv", index_col="Unnamed: 0")

pc5 = pd.read_csv(r"./../data/product_categories/product_category_3033-5857.csv", index_col="Unnamed: 0")
pc6 = pd.read_csv(r"./../data/product_categories/product_category_5858-6889.csv", index_col="Unnamed: 0")
pc7 = pd.read_csv(r"./../data/product_categories/product_category_6890-7316.csv", index_col="Unnamed: 0")
pc8 = pd.read_csv(r"./../data/product_categories/product_category_7317-7554.csv", index_col="Unnamed: 0")
pc9 = pd.read_csv(r"./../data/product_categories/product_category_7555-7906.csv", index_col="Unnamed: 0")
pc10 = pd.read_csv(r"./../data/product_categories/product_category_7907-8055.csv", index_col="Unnamed: 0")
pc11 = pd.read_csv(r"./../data/product_categories/product_category_8056-8316.csv", index_col="Unnamed: 0")
pc12 = pd.read_csv(r"./../data/product_categories/product_category_8317-8734.csv", index_col="Unnamed: 0")

pc1

pc2.index = range(716, 1206)

pc2

pc3

pc4

pc5

pc6

pc7

pc8

pc9

pc10

pc11

pc12

product_categories = pd.concat([pc1, pc2, pc3, pc4, pc5, pc6, pc7, pc8, pc9, pc10, pc11, pc12])

all_collection_categories = pd.concat([all, product_categories], axis=1)

all_collection_categories.to_csv('all_categories.csv', index=False)

## CHECKING FOR DUPLICATES AGAIN

all_collection_categories.duplicated().sum()

len(all_collection_categories)

all_collection_categories.duplicated(["link", "img"]).sum()

no_duplicates = all_collection_categories.drop_duplicates(subset=["link", "img"])

len(no_duplicates)

no_duplicates.reset_index(inplace=True, drop=True)

# transform values to lowercase
no_duplicates.product_category = no_duplicates.product_category.str.lower()

# save new dataframe to csv
no_duplicates.to_csv("all_categories_unqiue.csv", index=False)
