==> python_codes/4.2_generalize_categories.ipynb <==
# Create new category column (generalize categories taken from web scraping)

## IMPORT LIBRARIES

import pandas as pd
import re
import numpy as np

## LOAD DATA

all = pd.read_csv(r'./../data/cleaned_data/all_categories_unqiue.csv')

all.head()

## PREPARING GENERALIZED LABELS

# check all categories taken from the website
all.product_category.unique()

# lists of keywords to look for in the product_category column to apply different labels
sports_accessories = [r"mat[s]{0,1}", r"sport[s]{0,1}"]
hair_accessories = [r"clip[s]{0,1}", r'headband[s]{0,1}']
wallets = [r"leather goods", r"case[s]{0,1}"]
pouches = [r"pouch[es]{0,1}", r'toilet[s]{0,1}',]
outdoor_toys = [r"skateboard[s]{0,1}", r"scooter[s]{0,1}"]
umbrellas = [r"umbrella[s]{0,1}"]

swimsuits = [r"bikini[s]{0,1}", r"swimsuit[s]{0,1}", r"uv", r"swim"]
underwears = [r"boxer[s]{0,1}", r"brief[s]{0,1}",  r"bra[s]{0,1}", r"knicker[s]{0,1}", r"culotte[s]{0,1}", r"tight[s]{0,1}", r"underwear[s]{0,1}"]
sleepwear = [r"pyjama[s]{0,1}", r"kimono[s]{0,1}", r"nightdress[es]{0,1}", r"gown[s]{0,1}", r"robe[s]{0,1}"]

tops = [r"top[s]{0,1}", r"blouse[s]{0,1}", r"[t-]{0,1}shirt[s]{0,1}"]
jumpers = [r"jumper[s]{0,1}", r"cardigan[s]{0,1}",  r"sweatshirt[s]{0,1}", r"sweater[s]{0,1}", r"pull[s]{0,1}", r"hood[s]{0,1}"]
bottoms = [r"bottom[s]{0,1}", r"trouser[s]{0,1}", r"legging[s]{0,1}", r"shorts", r"pant[s]{0,1}", r"jean[s]{0,1}", r'skirt[s]{0,1}', r'chino[s]{0,1}', r'flare[s]{0,1}', r'slim[s]{0,1}', r'loose[s]{0,1}', r'tapered', r'bottom[s]{0,1}', r"jogger[s]{0,1}", r'straight[s]{0,1}', r"boyfriend[s]{0,1}"]
dress_overalls = [r"dress[es]{0,1}", r"dungaree[s]{0,1}", r"bodysuit[s]{0,1}", r"playsuit[s]{0,1}", r"set[s]{0,1}", r"jumpsuit[s]{0,1}"]
coats = [r"capes", r"ponchos", r"coat[s]{0,1}", r"jacket[s]{0,1}", r"parka[s]{0,1}"]

jewelries = [r'necklace[s]{0,1}', r"earring[s]{0,1}", r"bracelet[s]{0,1}", r"ring[s]{0,1}", r"watch[es]{0,1}", r"jewellery"]
accessories = [r"caps", r"[sun]{0,1}hat[s]{0,1}", r"beanie[s]{0,1}", r"belt[s]{0,1}", r"[sun]{0,1}glass[es]{0,1}", r"scarf", r"scarves", r"accessories", r"foulard[s]{0,1}", r"snood[s]{0,1}", r"glove[s]{0,1}", r"mitten[s]{0,1}", r"sock[s]{0,1}"]
bags = [r"bag[s]{0,1}", r"[back]{0,1}pack[s]{0,1}"]

> **NOTE: Although everything is being marked for their categories, we will only keep products with the following results:**
> | category     |label|
> |----------------|-----|
> | accessories    | 0 |
> | bag            | 1 |
> | bottom         | 2 |
> | coat           | 3 |
> | dress_overalls | 4 |
> | jewelry        | 5 |
> | jumper         | 6 |
> | top            | 7 |

# function to extract labels
def extract_label(df):
    """This function will use the lists provided above to extract the category of a product using the product column of the dataframe.
    Depending on the percentage of the data lost during the process, we will either review the list or proceed with the dataset acquired."""
    
    categories = [sports_accessories, hair_accessories, wallets,  pouches, outdoor_toys, umbrellas, swimsuits, underwears, sleepwear, 
                accessories, dress_overalls, bottoms, jewelries, jumpers, tops,  coats, bags]
    labels = ["sportswear", "hair_accessories", "wallet", "pouch", "scooter", "umbrella", "swimwear", "underwear", "sleepwear", 
              "accessories", "dress_overalls", "bottom", "jewelry", "jumper", "top", "coat", "bag"]
    
    # empty lists to store the data
    index = []
    category = []
    
    for ind in df.index:
        found=False
        for c, l in zip(categories, labels):
            for r in c:
                if re.search(r, df.product_category[ind]):
                    index.append(ind)
                    category.append(l)
                    found = True 
                    break # move on to the next row and avoid duplicates
                
                elif re.search(r, df['product'][ind]):
                    index.append(ind)
                    category.append(l)
                    found = True 
                    break # move on to the next row and avoid duplicates
                        
            if found: # check if flag is True
                break
            
        if not found:
            index.append(ind)
            category.append(np.nan) # fill missing labels with NaN

    # Create new label column 
    new_df = pd.DataFrame(category, index=index, columns=['category'])
    df_add = pd.concat([df, new_df], axis=1)
    missing = df_add[df_add.category.isna()==True]
    print(round((len(missing)/len(df))*100, 2), "% missing labels :",len(missing))
    
    return new_df, df_add, missing

new_df, all1, missing = extract_label(all)

missing

> **NOTE:** Since the number of missing values is minimal and most of them are not in the categories we wish to keep for the recommender, we will opt to drop them.

all2 = all1.dropna().reset_index(drop=True)

# save to csv
all2.to_csv('all_labelled_complete.csv', index=False)

# list of categories to keep
categories_to_keep = ["jumper", "top", "bottom", "dress_overalls", "coat", "jewelry", "accessories", "bag"]

# new df to take only items in the categories above
all_filtered = all2[all2['category'].isin(categories_to_keep)]

# double-checking unique values in the category column
all_filtered.category.unique() # seems to be in order

all_filtered.category.value_counts()

# reset index
all_filtered.reset_index(drop=True, inplace=True)

all_filtered.head()

# save to csv
all_filtered.to_csv('filtered_all.csv', index=False)

all_filtered.iloc[2199]

all_filtered.iloc[4979]
