==> python_codes/7_replacing_special_characters.ipynb <==
# Replacing Special Characters
- This step has been done in order to load the csv in SQL Workbench. Unfortunately, it is not able to read the file with accented letters.

## IMPORT LIBRARIES

import pandas as pd

all = pd.read_csv(r'./../data/cleaned_data/filtered_all.csv')

def change_char(df):
    """Function to replace special characters with normal alphabet"""
    special_characters = ["[àâä]","[éèêë]", "ï", "[öòóô]", "[ùûü]", "[²]", "[’]", "[°]"]
    replace = ["a", "e", "i", "o", "u", "2", "'", ""]

    for s, r in zip(special_characters, replace):
        for column in df.columns:
            df[column] = df[column].replace(s, r, regex=True)
    
    return df
        

all_new = change_char(all)

all_new

# save with encoding ascii required by sql and ignore any other errors from other special characters
all_new.to_csv("sql_csv.csv", index=False, encoding="ascii", errors="ignore")
